Mejoras por rondas y dificultades implementadas
Falta testearrlo para ver si funciona correctamente

+++ Bowling-Frenzy/Assets/Scripts/Enemies/BoloEBoos.cs
@@ -45,7 +45,7 @@ public class BoloEBoos : EnemyBase
         ApproachPlayer();
     }
 
-    public override void ReceiveDamage(int damage)
+    public override void ReceiveDamage(float damage)
     {
         health -= damage;
